requirements.txt
fastapi
uvicorn
pandas
git+https://github.com/rongardF/tvdatafeed.git
beautifulsoup4
requests
notion-client
pydantic-settings>=2.0.0
web3
google-cloud-bigquery>=3.11.4
google-auth>=2.17.3
sqlalchemy
psycopg2-binary
python-dotenv